feat(corewar): init vm and load champions into memory

+++ corewar/includes/config.h
@@ -1,6 +1,8 @@
 #ifndef CONFIG_H
 # define CONFIG_H
 
+# include <stddef.h>
+
 extern const int CHAMP_MAX_COUNT;
 extern const int REG_NUMBER;
 extern const int REG_SIZE;
@@ -8,7 +10,7 @@ extern const int CYCLE_TO_DIE;
 extern const int MAX_CHECKS;
 extern const int NBR_LIVE;
 extern const int CYCLE_DELTA;
-extern const int MEM_SIZE;
-extern const int CHAMP_MAX_SIZE;
+extern const size_t MEM_SIZE;
+extern const size_t CHAMP_MAX_SIZE;
 
 #endif
